refactor: Move formFieldProps to src/utils/props directory
