+++ src/components/nSelect/logic/nuiHandler.js
@@ -1,5 +1,5 @@
 import uuidMixin from '@/utils/uuid'
-import formField from '@/utils/formField/helpers/formFieldProps'
+import formField from '@/utils/props/formFieldProps'
 import nValidationAlerts from '@/utils/components/nValidationAlerts.vue'
 
 export default {
